refactor: Implement actionRunner for centralized error handling in ERD commands; refactor error handling logic from erdCommand
